requirements.txt
python-dotenv==1.0.0
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
PyJWT==2.8.0
bcrypt==4.0.1
passlib==1.7.4